// src/schemas/feedbackSchema.ts
import { z } from 'zod';
import { extendZod, zId } from "@zodyac/zod-mongoose";

// Extend zod with the zId and zUUID functions (!!important)
extendZod(z); 

// Define the schema (!!important)
export const FeedbackSchema = z.object({
    customer_id: zId("Customer"),
    product_id: zId("Product"),
    rating: z.number().min(0).max(5).default(0),
    comment: z.string().min(3).max(2000).default("test"),
    is_active: z.boolean().default(true),
    //default value will be generated by mongoose
    createdAt: z.date().default(new Date()), 
    updatedAt: z.date().default(new Date()),
});


// Define the schema with the ID (!!important)
export const zFeedbackSchemaUdate = FeedbackSchema.extend({
    _id: zId(), //default value will be generated by mongoose
    __v: z.number().optional()
})

export type Feedback = z.infer<typeof zFeedbackSchemaUdate>;